fixed mysterious no-printf breaks things bug

+++ c-revamp/spf.h
@@ -198,8 +198,8 @@ class SPF {
     
         // parameter updates
         void update_shape(int user, int item, int rating);
-        float update_tau(int user);
-        float update_theta(int user);
+        double update_tau(int user);
+        double update_theta(int user);
         void update_beta(int item);
 
         double get_ave_log_likelihood();
